requirements.txt
-i  https://pypi.tuna.tsinghua.edu.cn/simple
Pillow==9.4.0
requests==2.28.2
numpy==1.24.3
pyyaml==6.0